test(e2e): harness-ui — absorb 6 standalone probes (app-icon-present skipLaunch + 5 store-driven)
Bucket-1 absorption (task #222) on top of #223's per-case capability surface.

Absorbed:
- app-icon-present (skipLaunch: true — pure fs/json check, no electron)
- group-add (per-group + button creates session in clicked group)
- import-empty-groups (importSession synthesizes default group)
- rename (InlineRename: Enter / Escape / whitespace / click-outside / IME)
- terminal (Bash tool block expand → xterm host renders ANSI)
- tool-render-open-in-editor (long stdout button + IPC noop write — uses
  spec.launch.env to set CCSM_OPEN_IN_EDITOR_NOOP=1; harmless to other cases)

Reclassified to bucket 4 (kept as standalone probe-e2e-dnd):
- dnd: needs CCSM_E2E_HIDDEN=0 (visible window) for dnd-kit pointer
  hit-testing. The capability surface has no per-case env override and
  flipping the whole harness to visible would slow every other case +
  introduce window pop-up noise on local runs.

Verification:
- node scripts/harness-ui.mjs → 28/28 PASS (was 22/22 + 6 new = 28).
- Each absorbed case also passes in isolation via --only.
- docs/e2e-runner.md "must stay" list updated; absorbed entries removed,
  dnd reason documented.
